+++ osh/history.py
@@ -18,7 +18,7 @@ from osh import word_
 from typing import Optional, TYPE_CHECKING
 if TYPE_CHECKING:
   from frontend.parse_lib import ParseContext
-  from core.pyutil import Readline
+  from frontend.py_readline import Readline
   from core.util import _DebugFile
